BuiQuangTruong - update auto test code base (#8)
* update code base automation test

* resolve comment
