+++ AutomationTest/.env
@@ -1,0 +1,3 @@
+BASE_LOCAL_URL=""
+USER_NAME=""
+PASSWORD=""
