+++ AutomationTest/playwright.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
   workers: process.env.CI ? 1 : undefined,
   reporter: [['junit', { outputFile: 'test-results/e2e-junit-results.xml' }]],
   use: {
-    baseURL: 'https://connectingmentor.netlify.app',
+    baseURL: process.env.BASE_LOCAL_URL,
     trace: 'retain-on-failure',
     headless: false,
     testIdAttribute: '',
