--- AutomationTest/src/data-types/user.ts
@@ -1,4 +1,0 @@
-export interface LoginUser {
-    email: string ;
-    password: string;
-  }
