+++ AutomationTest/src/models/user/user.ts
@@ -1,0 +1,4 @@
+export interface LoginUser {
+  email: string;
+  password: string;
+}
